bidder dash working
