+++ src/components/ItemForm.js
@@ -5,6 +5,35 @@ import axios from 'axios';
 import styled from 'styled-components';
 
 // start of styled components
+const Img = styled.img `
+    width: 200px;
+    height: 240px;
+    display: block;
+    margin: auto;
+    border: 1px solid gray;
+`
+
+const FileSelect = styled.div`    
+    width: 350px;
+    margin-top: 4% 0;
+    margin-left: 9%;
+    padding: 4%;   
+`
+const UploadButton = styled.button`
+    width: 20%;
+    margin-left: -22%
+    margin-right: 12%
+`
+const ItemBox = styled.div `
+overflow: auto;
+border: 1px solid grey;
+width: 200px;
+height: 250px;
+padding: 4%;
+margin-left: 22%;
+border-radius: 10px;
+box-shadow: 5px 5px 10px rgba(0, 0, 0, 0.4);
+`
 
 const Label = styled.label `
     margin-top: -10%;
@@ -52,7 +81,29 @@ useEffect(() => {
     status && setName(nameForm => [...nameForm, status])
 }, [status]);
 // Button event handler
+const [selectedFile, setSelectedFile] = useState([]);
+const fileSelectedHandler = event => {
+    setSelectedFile(URL.createObjectURL(event.target.files[0]));
+}
 
+// 
+
+    const fileUploadHandler = () => {
+            //form data being sent  
+        const frmdata = new FormData();
+            frmdata.append( selectedFile, selectedFile.name);
+        axios
+        .post('silent-auction-be.herokuapp.com/:sellerId/items/', frmdata, {
+            onUploadProgress: progressEvent => {
+            console.log('Upload Progress: ' + (progressEvent.loaded / progressEvent.total * 100) + '%')}
+            })
+        .then(res => {
+            console.log(res);
+           
+            });
+      
+           
+            }
 
 return (
   
@@ -121,7 +172,22 @@ return (
         <p className='errors' >{errors.description}</p>
         )}
         
+{/*  paste add item so uploading all */}
+                    <UploadButton 
+                    onClick={fileUploadHandler}            
+                    >Add Item</UploadButton>
 
+                    <FileSelect>
+                    <input className='choose-file'      
+                    type='file'          
+                    onChange={fileSelectedHandler}
+                    /> 
+                                    
+
+                    </FileSelect>
+                    <ItemBox>
+                    <Img src={selectedFile} alt='item'/>
+                    </ItemBox>
     </Form>
